+++ cardigan/.storybook/preview.js
@@ -49,7 +49,8 @@ export const parameters = {
       order: [
         'Cardigan',
         'Components',
-        ['Banners', 'Buttons', ['Basics', 'Alternates'], 'Cards'],
+        ['Banners', 'Buttons', ['Basics', 'Alternates'], 'Cards', 'Images', 'Inputs', 'Links', 'Media', 'Tags'],
+        'To be made reusable',
 
         'Global',
       ],
