+++ cardigan/stories/components/Images/CatalogueImageGallery/mock-images.ts
